+++ packages/carbon-graphs/src/js/controls/Scatter/helpers/helpers.js
@@ -156,9 +156,10 @@ const getDataPointValues = (target) => target.internalValuesSubset;
  * @param {object} scale - d3 scale for Graph
  * @param {object} config - Graph config object derived from input JSON
  * @param {Array} pointGroupPath - d3 html element of the points group
+ * @param {object} legendSVG - d3 element path of the legend from the parent control
  * @returns {undefined} - returns nothing
  */
-const drawDataPoints = (scale, config, pointGroupPath) => {
+const drawDataPoints = (scale, config, pointGroupPath, legendSVG) => {
   const renderDataPointPath = (path, value, index) => path.append(() => new Shape(getShapeForTarget(value)).getShapeElement(
     getDefaultSVGProps({
       svgClassNames: styles.point,
@@ -168,12 +169,7 @@ const drawDataPoints = (scale, config, pointGroupPath) => {
         dataPointActionHandler(value, index, this);
       },
       a11yAttributes: {
-        'aria-hidden':
-          document
-            .querySelector(
-              `.${styles.legendItem}[aria-describedby="${value.key}"]`,
-            )
-            ?.getAttribute('aria-current') === 'false',
+        'aria-hidden': legendSVG ? legendSVG.select(`.${styles.legendItem}[aria-describedby='${value.key}']`)?.attr('aria-current') === 'false' : 'false',
         'aria-describedby': value.key,
         'aria-disabled': !utils.isFunction(value.onClick),
       },
@@ -254,9 +250,10 @@ const drawDataPoints = (scale, config, pointGroupPath) => {
  * @param {object} config - config object derived from input JSON
  * @param {d3.selection} canvasSVG - d3 selection node of canvas svg
  * @param {Array} dataTarget - Data points
+ * @param {object} legendSVG - d3 element path of the legend from the parent control
  * @returns {undefined} - returns nothing
  */
-const draw = (scale, config, canvasSVG, dataTarget) => {
+const draw = (scale, config, canvasSVG, dataTarget, legendSVG) => {
   const scatterSVG = canvasSVG
     .append('g')
     .classed(styles.scatterGraphContent, true)
@@ -285,7 +282,7 @@ const draw = (scale, config, canvasSVG, dataTarget) => {
     .select(`.${styles.currentPointsGroup}`)
     .selectAll(`.${styles.point}`)
     .data(getDataPointValues(dataTarget).filter((d) => d.y !== null));
-  drawDataPoints(scale, config, pointPath.enter());
+  drawDataPoints(scale, config, pointPath.enter(), legendSVG);
   pointPath
     .exit()
     .transition()