+++ packages/carbon-graphs/CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 * Changed
+  * For consistency replaced `document` in all draw helpers with `graph.legendSVG` for better control.
+
+* Changed
   * For consistency updated reflow in Graph and Gantt constructs to update the eventlines.
   * Added code to handle null/undefined/blank in paired result graph during both initial load and reflow.